move load_clientconfig from zingolib into zingoconfig
